requirements.txt
django
djangorestframework
djangorestframework-jsonapi

psycopg2
